Move DTOs to own package
