+++ app/src/test/kotlin/no/nav/tiltakspenger/vedtak/person/PersonRoutesTest.kt
@@ -14,19 +14,13 @@ import io.ktor.client.statement.bodyAsText
 import io.ktor.http.HttpHeaders
 import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
-import io.ktor.server.application.install
-import io.ktor.server.auth.Authentication
-import io.ktor.server.auth.jwt.JWTPrincipal
-import io.ktor.server.auth.jwt.jwt
 import io.ktor.server.testing.ApplicationTestBuilder
 import io.ktor.server.testing.testApplication
-import no.nav.tiltakspenger.vedtak.routes.auth
 import no.nav.tiltakspenger.vedtak.routes.jacksonSerialization
 import no.nav.tiltakspenger.vedtak.routes.openAPI
-import no.nav.tiltakspenger.vedtak.routes.person.PersonDTO
+import no.nav.tiltakspenger.vedtak.modell.PersonDTO
 import no.nav.tiltakspenger.vedtak.routes.person.person
 import no.nav.tiltakspenger.vedtak.routes.person.personPath
-import no.nav.tiltakspenger.vedtak.routes.person.personRoutes
 import org.junit.jupiter.api.Test
 import org.skyscreamer.jsonassert.JSONAssert
 import org.skyscreamer.jsonassert.JSONCompareMode
